# userprofiledataservAPI/bundled-openapi.yaml
openapi: 3.0.3
info:
  title: User Management API
  description: |
    A comprehensive multi-tenant user identity and profile management REST API. This API provides 
    functionality for managing user identities, authentication, authorization, and profile data 
    across multiple tenants.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
    url: 'https://example.com/support'
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'https://api.example.com/v1'
    description: Production server
  - url: 'https://staging-api.example.com/v1'
    description: Staging server
  - url: 'https://dev-api.example.com/v1'
    description: Development server
...
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: List users
      description: Get a paginated list of users within the tenant
      operationId: listUsers
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - $ref: '#/components/parameters/ApiVersionHeader'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationOffset'
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Search term to filter users by name or email
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - active
              - inactive
              - suspended
              - pending
          description: Filter users by status
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            enum:
              - email
              - firstName
              - lastName
              - status
              - createdAt
            default: createdAt
          description: Field to sort the results by
        - name: sortOrder
          in: query
          required: false
          schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
          description: Sort order (ascending or descending)
      responses:
        '200':
          description: List of users
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
    post:
      tags:
        - Users
      summary: Create a user
      description: Create a new user within the tenant
      operationId: createUser
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - $ref: '#/components/parameters/ApiVersionHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created successfully
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            IdempotencyRequestId:
              $ref: '#/components/headers/IdempotencyRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Conflict - User with this email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
  parameters:
    TenantIdHeader:
      name: X-Tenant-ID
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: ID of the tenant to operate within
    ApiVersionHeader:
      name: X-API-Version
      in: header
      required: false
      schema:
        type: string
        default: 1.0.0
      description: API version requested by the client
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: Client-provided idempotency key for safe retries of non-idempotent operations
...
    PaginationLimit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Maximum number of items to return
    PaginationOffset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Number of items to skip
  headers:
    ETag:
      description: Entity tag for the resource version
      schema:
        type: string
    IdempotencyRequestId:
      description: Unique identifier of the original request that was processed
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Authentication credentials are missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: The authenticated user does not have permission to access the requested resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ValidationError:
      description: The request contains invalid data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Error message
        details:
          type: object
          description: Additional error details
    ValidationError:
      type: object
      required:
        - code
        - message
        - fields
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Error message
        fields:
          type: array
          items:
            type: object
            required:
              - field
              - message
            properties:
              field:
                type: string
                description: Name of the field with validation error
              message:
                type: string
                description: Validation error message
    User:
      type: object
      required:
        - id
        - email
        - firstName
        - lastName
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the user
        email:
          type: string
          format: email
          description: User's email address
        firstName:
          type: string
          description: User's first name
        lastName:
          type: string
          description: User's last name
        status:
          type: string
          enum:
            - active
            - inactive
            - suspended
            - pending
          description: Current status of the user
        phoneNumber:
          type: string
          description: User's phone number
        lastLoginAt:
          type: string
          format: date-time
          description: Timestamp of the user's last login
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the user was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the user was last updated
        deletedAt:
          type: string
          format: date-time
          description: Timestamp when the user was soft-deleted (null if not deleted)
        tenantId:
          type: string
          format: uuid
          description: ID of the tenant the user belongs to
    UserCreate:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
        - tenantId
      properties:
        email:
          type: string
          format: email
          description: User's email address
        password:
          type: string
          format: password
          minLength: 8
          description: User's password
        firstName:
          type: string
          description: User's first name
        lastName:
          type: string
          description: User's last name
        phoneNumber:
          type: string
          description: User's phone number
        tenantId:
          type: string
          format: uuid
          description: ID of the tenant the user belongs to
        roleIds:
          type: array
          items:
            type: string
            format: uuid
          description: IDs of roles to assign to the user
...
    UserList:
      type: object
      required:
        - users
        - totalCount
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
        totalCount:
          type: integer
          description: Total number of users matching the query
        limit:
          type: integer
          description: Maximum number of items returned
        offset:
          type: integer
          description: Number of items skipped
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []